Split write to cache testing into individual files so that it can be run with "npm run test:unit"
