+++ test/unit/operations/write/writeOnlyValuesToARoot.ts
@@ -1,0 +1,45 @@
+import { CacheContext } from '../../../../src/context';
+import { GraphSnapshot } from '../../../../src/GraphSnapshot';
+import { EntitySnapshot } from '../../../../src/nodes';
+import { write } from '../../../../src/operations/write';
+import { NodeId, StaticNodeId } from '../../../../src/schema';
+import { query, strictConfig } from '../../../helpers';
+
+const { QueryRoot: QueryRootId } = StaticNodeId;
+
+// These are really more like integration tests, given the underlying machinery.
+//
+// It just isn't very fruitful to unit test the individual steps of the write
+// workflow in isolation, given the contextual state that must be passed around.
+describe(`operations.write`, () => {
+
+  const context = new CacheContext(strictConfig);
+  const empty = new GraphSnapshot();
+  const rootValuesQuery = query(`{ foo bar }`);
+
+  describe(`write only values to a root`, () => {
+    let snapshot: GraphSnapshot, editedNodeIds: Set<NodeId>;
+    beforeAll(() => {
+      const result = write(context, empty, rootValuesQuery, { foo: 123, bar: 'asdf' });
+      snapshot = result.snapshot;
+      editedNodeIds = result.editedNodeIds;
+    });
+
+    it(`creates the query root, with the values`, () => {
+      expect(snapshot.get(QueryRootId)).to.deep.eq({ foo: 123, bar: 'asdf' });
+    });
+
+    it(`marks the root as edited`, () => {
+      expect(Array.from(editedNodeIds)).to.have.members([QueryRootId]);
+    });
+
+    it(`only contains the root node`, () => {
+      expect(snapshot.allNodeIds()).to.have.members([QueryRootId]);
+    });
+
+    it(`emits the root as an EntitySnapshot`, () => {
+      expect(snapshot.getNodeSnapshot(QueryRootId)).to.be.an.instanceOf(EntitySnapshot);
+    });
+  });
+
+});
